connection and app separation
